[PM-15052] Refactored: Moved revoke user methods into a separate command class.
